重命名renders。解决2dimage颜色多一倍的问题，原因是使用帧缓存的情况下，使用GL_ONE Blend了两次，导致颜色值翻倍

+++ coding/gamemachine/src/gmdatacore/imagereader/gmimagereader_bmp.cpp
@@ -126,7 +126,7 @@ void GMImageReader_BMP::writeDataToImage(BitmapFile& bitmap, GMImage* img, GMuin
 #if USE_OPENGL
 	data.target = GL_TEXTURE_2D;
 	data.mipLevels = 1;
-	data.internalFormat = GL_RGB16;
+	data.internalFormat = GL_RGB8;
 	data.format = GL_BGR;
 	data.swizzle[0] = GL_RED;
 	data.swizzle[1] = GL_GREEN;
